+++ packages/graph/hash_graph/bench/benches/representative_read/ontology/entity_type.rs
@@ -1,6 +1,6 @@
 use criterion::{BatchSize::SmallInput, Bencher};
 use graph::{
-    store::{query::Expression, EntityTypeStore},
+    store::{query::Filter, EntityTypeStore},
     subgraph::{GraphResolveDepths, StructuralQuery},
 };
 use rand::{prelude::IteratorRandom, thread_rng};
@@ -27,7 +27,7 @@ pub fn bench_get_entity_type_by_id(
         |entity_type_id| async move {
             store
                 .get_entity_type(&StructuralQuery {
-                    expression: Expression::for_versioned_uri(&entity_type_id),
+                    filter: Filter::for_versioned_uri(&entity_type_id),
                     graph_resolve_depths: GraphResolveDepths {
                         data_type_resolve_depth: 0,
                         property_type_resolve_depth: 0,