+++ packages/graph/hash_graph/lib/graph/src/shared/mod.rs
@@ -1,3 +1,4 @@
 pub mod identifier;
 pub mod provenance;
+pub mod query;
 pub mod subgraph;