+++ packages/graph/hash_graph/lib/graph/src/store/postgres/query/expression/with_clause.rs
@@ -62,6 +62,7 @@ mod tests {
     use crate::{
         ontology::DataTypeQueryPath,
         store::postgres::query::{
+            expression::OrderByExpression,
             test_helper::{max_version_expression, trim_whitespace},
             Expression, Path, SelectExpression, SelectStatement, Table, TableName, WhereExpression,
         },
@@ -74,7 +75,7 @@ mod tests {
 
         with_clause.add_statement(TableName::TypeIds, SelectStatement {
             with: WithExpression::default(),
-            distinct: false,
+            distinct: Vec::new(),
             selects: vec![
                 SelectExpression::new(Expression::Asterisk, None),
                 SelectExpression::new(
@@ -88,6 +89,7 @@ mod tests {
             },
             joins: vec![],
             where_expression: WhereExpression::default(),
+            order_by_expression: OrderByExpression::default(),
         });
 
         assert_eq!(
@@ -100,7 +102,7 @@ mod tests {
 
         with_clause.add_statement(TableName::DataTypes, SelectStatement {
             with: WithExpression::default(),
-            distinct: false,
+            distinct: Vec::new(),
             selects: vec![SelectExpression::new(Expression::Asterisk, None)],
             from: Table {
                 name: TableName::DataTypes,
@@ -108,6 +110,7 @@ mod tests {
             },
             joins: vec![],
             where_expression: WhereExpression::default(),
+            order_by_expression: OrderByExpression::default(),
         });
 
         assert_eq!(