+++ packages/graph/hash_graph/lib/graph/src/store/postgres/query/statement/select.rs
@@ -1,17 +1,25 @@
 use std::fmt::{self, Write};
 
 use crate::store::postgres::query::{
-    JoinExpression, SelectExpression, Table, Transpile, WhereExpression, WithExpression,
+    expression::OrderByExpression, Column, JoinExpression, SelectExpression, Table, Transpile,
+    WhereExpression, WithExpression,
 };
 
 #[derive(Debug, PartialEq, Eq, Hash)]
 pub struct SelectStatement<'q> {
     pub with: WithExpression<'q>,
-    pub distinct: bool,
+    pub distinct: Vec<Column<'q>>,
     pub selects: Vec<SelectExpression<'q>>,
     pub from: Table,
     pub joins: Vec<JoinExpression<'q>>,
     pub where_expression: WhereExpression<'q>,
+    pub order_by_expression: OrderByExpression<'q>,
+}
+
+#[derive(Debug, Copy, Clone, PartialEq, Eq, Hash)]
+pub enum Distinctness {
+    Indestinct,
+    Destinct,
 }
 
 impl Transpile for SelectStatement<'_> {
@@ -23,8 +31,16 @@ impl Transpile for SelectStatement<'_> {
 
         fmt.write_str("SELECT ")?;
 
-        if self.distinct {
-            fmt.write_str("DISTINCT ")?;
+        if !self.distinct.is_empty() {
+            fmt.write_str("DISTINCT ON(")?;
+
+            for (idx, column) in self.distinct.iter().enumerate() {
+                column.transpile(fmt)?;
+                if idx + 1 < self.distinct.len() {
+                    fmt.write_str(", ")?;
+                }
+            }
+            fmt.write_str(") ")?;
         }
 
         for (idx, condition) in self.selects.iter().enumerate() {
@@ -46,6 +62,11 @@ impl Transpile for SelectStatement<'_> {
             self.where_expression.transpile(fmt)?;
         }
 
+        if !self.order_by_expression.is_empty() {
+            fmt.write_char('\n')?;
+            self.order_by_expression.transpile(fmt)?;
+        }
+
         Ok(())
     }
 }
@@ -56,6 +77,7 @@ mod tests {
 
     use postgres_types::ToSql;
     use type_system::{DataType, EntityType, PropertyType};
+    use uuid::Uuid;
 
     use crate::{
         knowledge::{Entity, EntityQueryPath, Link, LinkQueryPath},
@@ -63,7 +85,10 @@ mod tests {
             DataTypeQueryPath, EntityTypeQueryPath, LinkTypeQueryPath, PropertyTypeQueryPath,
         },
         store::{
-            postgres::query::{test_helper::trim_whitespace, PostgresQueryRecord, SelectCompiler},
+            postgres::query::{
+                test_helper::trim_whitespace, Distinctness, Ordering, PostgresQueryRecord,
+                SelectCompiler,
+            },
             query::{Filter, FilterExpression, Parameter},
         },
     };
@@ -188,7 +213,7 @@ mod tests {
             &compiler,
             r#"
             WITH "type_ids" AS (SELECT *, MAX("type_ids"."version") OVER (PARTITION BY "type_ids"."base_uri") AS "latest_version" FROM "type_ids")
-            SELECT DISTINCT *
+            SELECT *
             FROM "data_types"
             JOIN "type_ids" AS "type_ids_0_0"
               ON "type_ids_0_0"."version_id" = "data_types"."version_id"
@@ -213,7 +238,7 @@ mod tests {
             &compiler,
             r#"
             WITH "type_ids" AS (SELECT *, MAX("type_ids"."version") OVER (PARTITION BY "type_ids"."base_uri") AS "latest_version" FROM "type_ids")
-            SELECT DISTINCT *
+            SELECT *
             FROM "data_types"
             JOIN "type_ids" AS "type_ids_0_0"
               ON "type_ids_0_0"."version_id" = "data_types"."version_id"
@@ -274,7 +299,8 @@ mod tests {
             FROM "property_types"
             JOIN "property_type_data_type_references" AS "property_type_data_type_references_0_0"
               ON "property_type_data_type_references_0_0"."source_property_type_version_id" = "property_types"."version_id"
-            JOIN "data_types" AS "data_types_0_1" ON "data_types_0_1"."version_id" = "property_type_data_type_references_0_0"."target_data_type_version_id"
+            JOIN "data_types" AS "data_types_0_1"
+              ON "data_types_0_1"."version_id" = "property_type_data_type_references_0_0"."target_data_type_version_id"
             JOIN "property_type_data_type_references" AS "property_type_data_type_references_1_0"
               ON "property_type_data_type_references_1_0"."source_property_type_version_id" = "property_types"."version_id"
             JOIN "type_ids" AS "type_ids_1_1"
@@ -396,7 +422,9 @@ mod tests {
                 "entities"."entity_id",
                 "entities"."version",
                 "entity_types_0_0"."schema"->>'$id',
-                "entities"."owned_by_id"
+                "entities"."owned_by_id",
+                "entities"."created_by_id",
+                "entities"."updated_by_id"
             FROM "entities"
             JOIN "entity_types" AS "entity_types_0_0"
               ON "entity_types_0_0"."version_id" = "entities"."entity_type_version_id"
@@ -422,18 +450,62 @@ mod tests {
             &compiler,
             r#"
             WITH "entities" AS (SELECT *, MAX("entities"."version") OVER (PARTITION BY "entities"."entity_id") AS "latest_version" FROM "entities")
-            SELECT DISTINCT
+            SELECT
                 "entities"."properties",
                 "entities"."entity_id",
                 "entities"."version",
                 "entity_types_0_0"."schema"->>'$id',
-                "entities"."owned_by_id"
+                "entities"."owned_by_id",
+                "entities"."created_by_id",
+                "entities"."updated_by_id"
             FROM "entities"
             JOIN "entity_types" AS "entity_types_0_0"
               ON "entity_types_0_0"."version_id" = "entities"."entity_type_version_id"
             WHERE "entities"."version" = "entities"."latest_version"
             "#,
             &[],
+        );
+    }
+
+    #[test]
+    fn entity_with_manual_selection() {
+        let mut compiler = SelectCompiler::<Entity>::new();
+        compiler.add_selection_path(
+            &EntityQueryPath::Id,
+            Distinctness::Destinct,
+            Some(Ordering::Ascending),
+        );
+        compiler.add_selection_path(
+            &EntityQueryPath::Version,
+            Distinctness::Destinct,
+            Some(Ordering::Descending),
+        );
+        compiler.add_selection_path(
+            &EntityQueryPath::Properties(None),
+            Distinctness::Indestinct,
+            None,
+        );
+
+        let filter = Filter::Equal(
+            Some(FilterExpression::Path(EntityQueryPath::CreatedById)),
+            Some(FilterExpression::Parameter(Parameter::Uuid(Uuid::nil()))),
+        );
+        compiler.add_filter(&filter);
+
+        test_compilation(
+            &compiler,
+            r#"
+            SELECT
+                DISTINCT ON("entities"."entity_id", "entities"."version")
+                "entities"."entity_id",
+                "entities"."version",
+                "entities"."properties"
+            FROM "entities"
+            WHERE "entities"."created_by_id" = $1
+            ORDER BY "entities"."entity_id" ASC,
+                     "entities"."version" DESC
+            "#,
+            &[&Uuid::nil()],
         );
     }