+++ packages/hash/api/src/model/knowledge/link.model.ts
@@ -1,5 +1,5 @@
 import {
-  StructuralQuery,
+  LinkStructuralQuery,
   GraphApi,
   PersistedLink,
 } from "@hashintel/hash-graph-client";
@@ -82,11 +82,11 @@ export default class {
 
   static async getByQuery(
     graphApi: GraphApi,
-    query: object,
-    options?: Omit<Partial<StructuralQuery>, "query">,
+    filter: object,
+    options?: Omit<Partial<LinkStructuralQuery>, "query">,
   ): Promise<LinkModel[]> {
     const { data: linkRootedSubgraphs } = await graphApi.getLinksByQuery({
-      query,
+      filter,
       graphResolveDepths: {
         dataTypeResolveDepth:
           options?.graphResolveDepths?.dataTypeResolveDepth ?? 0,
@@ -121,13 +121,17 @@ export default class {
 
     const linkModels = await LinkModel.getByQuery(graphApi, {
       all: [
-        { eq: [{ path: ["source", "id"] }, { literal: sourceEntityId }] },
-        { eq: [{ path: ["target", "id"] }, { literal: targetEntityId }] },
         {
-          eq: [
+          equal: [{ path: ["source", "id"] }, { parameter: sourceEntityId }],
+        },
+        {
+          equal: [{ path: ["target", "id"] }, { parameter: targetEntityId }],
+        },
+        {
+          equal: [
             { path: ["type", "versionedUri"] },
             {
-              literal: linkTypeId,
+              parameter: linkTypeId,
             },
           ],
         },