+++ packages/hash/frontend/src/graphql/queries/comment.queries.ts
@@ -28,6 +28,15 @@ export const createPersistedComment = gql`
   }
 `;
 
+export const resolvePersistedComment = gql`
+  mutation resolvePersistedComment($entityId: ID!) {
+    resolvePersistedComment(entityId: $entityId) {
+      ...CommentFields
+    }
+  }
+  ${commentFieldsFragment}
+`;
+
 export const deletePersistedComment = gql`
   mutation deletePersistedComment($entityId: ID!) {
     deletePersistedComment(entityId: $entityId) {